feat: 分离分析逻辑，正确显示日期范围
